refactor: update sample data name to `productData.mock`
