+++ src/components/ProductDetail/__mocks__/productData.mock.js
